abstract operations
