+++ src/utilities.py
@@ -3,6 +3,8 @@ from gi.repository import Gdk
 
 from graphs.data import Data
 
+import numpy
+
 
 def remove_unused_config_keys(config, template):
     delete_list = []
@@ -170,3 +172,28 @@ def swap(str1):
     str1 = str1.replace(".", ", ")
     str1 = str1.replace("third", ".")
     return str1
+
+
+def get_value_at_fraction(fraction, start, end):
+    """
+    Obtain the selected value of an axis given at which percentage (in terms of
+    fraction) of the length this axis is selected given the start and end range
+    of this axis
+    """
+    log_start = numpy.log10(start)
+    log_end = numpy.log10(end)
+    log_range = log_end - log_start
+    log_value = log_start + log_range * fraction
+    return pow(10, log_value)
+
+
+def get_fraction_at_value(value, start, end):
+    """
+    Obtain the fraction of the total length of the selected axis a specific
+    value corresponds to given the start and end range of the axis.
+    """
+    log_start = numpy.log10(start)
+    log_end = numpy.log10(end)
+    log_value = numpy.log10(value)
+    log_range = log_end - log_start
+    return (log_value - log_start) / log_range
